Extracción y almacenamiento correcto en bbdd.
