+++ app/src/main/java/es/meliseoperez/MainActivity.java
@@ -23,7 +23,7 @@ import java.util.concurrent.ExecutorService;
 
 import es.meliseoperez.safehaven.R;
 import es.meliseoperez.safehaven.api.aemet.AlertInfo;
-import es.meliseoperez.safehaven.api.aemet.AlertsExtractor2;
+import es.meliseoperez.safehaven.api.aemet.AlertsExtractor;
 import es.meliseoperez.safehaven.api.aemet.DownloadAndStoreJSONAlerts;
 import es.meliseoperez.safehaven.api.aemet.AlertsExtractor;
 import es.meliseoperez.safehaven.api.aemet.MyCallBack;
